feat: add CockroachDB v26 support alongside PostgreSQL
- Dialect detection and config (database.driver: cockroach|postgres)
- Lease-table migration lock replacing pg_try_advisory_lock on CRDB
- cockroach/ overlay directory with 23 variant SQL migration files
- Per-statement autocommit for CRDB migration execution
- Runtime advisory lock replacement via runtime_locks lease table
- READ COMMITTED isolation via per-connection SET on cockroachReadCommittedConnector
- GROUPING SETS rewrite to UNION ALL (CRDB does not support GROUPING SETS)
- Fix ctid -> id (PK) for batch deletes
- Fix pg_database -> SHOW DATABASES in setup auto-config for CRDB
- Fix percentile_cont ORDER BY int -> float8
- Fix LEAST untyped placeholders with explicit timestamptz casts
- Fix explicit NULLS LAST/FIRST for proxy expiry sort
- Fix interval type inference with int8 cast
- Fix float*int arithmetic with float8 casts
- Fix isUsageLogsPartitioned short-circuit on CRDB
- Fix ON CONFLICT COALESCE arbiter: DELETE+INSERT path for CRDB
- Standalone cmd/migrate command for migration-only deploys
- docker-compose.cockroach.yml for containerised CRDB deployment
- Integration test harness: INTEGRATION_DB_DSN/REDIS_ADDR/REDIS_PASSWORD overrides
- All 188 migrations verified on CockroachDB v26.2.2; integration suite green

+++ backend/internal/repository/usage_cleanup_repo.go
@@ -128,7 +128,7 @@ func (r *usageCleanupRepository) ClaimNextPendingTask(ctx context.Context, stale
 				OR (
 					status = $2
 					AND started_at IS NOT NULL
-					AND started_at < NOW() - ($3 * interval '1 second')
+					AND started_at < NOW() - ($3::int8 * interval '1 second')
 				)
 			ORDER BY created_at ASC
 			LIMIT 1
